Add files via upload
In this we need to connect the wires to the pin as instructed below:-
GPS module ==> Arduino

* GND ==> GND

* TX ==> Digital pin (D3)

* RX ==> Digital pin (D4)

*Vcc ==> 3.3 V

LCD==> Arduino * VSS ==> GND

* VCC ==> 5V

*VEE ==> 10K Resistor

*RS ==> A0 (Analog pin)

*R/W ==> GND

*E ==> A1

*D4 ==> A2

*D5 ==> A3

*D6 ==> A4

*D7 ==> A5

*LED+ ==> VCC

*LED- ==> GND


OUTPUT:-  After all the connections and uploading the code, the GPS module take some time to get satellite fix which is usually 15 to 20 minutes.If it takes more time go outdoor and try as it is not able to get the satellite fix inside the house. After that you can see that the lcd display can show the GPS cordinates.